Update test with the new paths and names
