+++ components/ir/connexion/form/mentions-legales/tests/MentionsLegales.test.jsx
@@ -1,17 +1,16 @@
 import { createShallow } from "@material-ui/core/test-utils";
 
-import { roles } from "../../config.json";
-import { RolesInput } from "../roles-input";
+import { MentionsLegales } from "../MentionsLegales";
 
-describe("components | connexion | roles-input", () => {
+describe("components | connexion | mentions-legales-text", () => {
   let shallow;
   beforeAll(() => {
     shallow = createShallow();
   });
   describe("snapshot", () => {
     it("doit correspondre avec les props requises", () => {
-      const props = { defaultValue: "depute", roles };
-      const wrapper = shallow(<RolesInput {...props} />);
+      const props = { classes: {} };
+      const wrapper = shallow(<MentionsLegales {...props} />);
       expect(wrapper).toBeDefined();
       expect(wrapper).toMatchSnapshot();
     });
